Comment on Instruction.js
I think it would be great if on a users first log in after signup the instruction card is the first thing that pops up. While adding gifs on the landing page is really helpful you can't guarantee users will pay attention to that. Maybe even if it is a condensed version of your current instruction modal, just something so uses can't miss it because I still feel myself a bit overwhelmed upon first logging in. Just a suggestion.